session works, check database
