+++ backend/.env
@@ -24,10 +24,11 @@ JWT_EXPIRATION_TIME="15m"
 # 42 intra
 INTRA_AUTH_URL="https://api.intra.42.fr/oauth/authorize"
 INTRA_TOKEN_URL="https://api.intra.42.fr/oauth/token"
-INTRA_APPLICATION_NAME="codam-aijox"
 INTRA_CLIENT_ID="a0da1686a7ee0637a1726a6fdab95af8d1fee124e617d4d7c808fe69fcef6df7"
 INTRA_CLIENT_SECRET="6cb49cd44805b99c8a9bec31431dba10478bfee0cde8d1081d1658a5d1793087"
 INTRA_CALLBACK_URL="http://localhost:3000/auth/redirect"
+INTRA_GET_ME_URL="https://api.intra.42.fr/v2/me"
+INTRA_APPLICATION_NAME="codam-aijox"
 
 # session
 SESSION_SECRET="lakshflkasfkljhlaskh"
