+++ backend/src/auth/auth.module.ts
@@ -5,12 +5,18 @@ import { OAuthStrategy } from './oauth/oauth.strategy';
 import { HttpModule } from '@nestjs/axios';
 import { JwtModule } from '@nestjs/jwt';
 import { UserModule } from 'src/user/user.module';
+import { SessionSerializer } from './utils/Serializer';
 
 @Module({
-  imports: [HttpModule, JwtModule.register({}), UserModule],
+  imports: [
+    HttpModule,
+    JwtModule.register({}),
+    UserModule
+  ],
   providers: [
     OAuthStrategy,
     AuthService,
+    SessionSerializer,
   ],
   controllers: [AuthController],
 })
