--- backend/src/auth/oauth/auth.ts
@@ -1,7 +1,0 @@
-import { UserDetails } from "../../user/utils.ts/types";
-
-export interface AuthenticationProvider {
-	validateUser(details: UserDetails);
-	createUser(details: UserDetails);
-	findUser(details: UserDetails);
-}
