+++ backend/src/auth/utils/types.ts
@@ -1,0 +1,3 @@
+import { User } from "../../user/user.entity"
+
+export type Done = (err: Error, user: User) => void;
