+++ backend/src/user/user.service.ts
@@ -1,9 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, getRepository } from 'typeorm';
+import { Repository } from 'typeorm';
 import User from './user.entity';
-import { CreateUserDto, ValidateUserDto } from './dto';
-import { UserDetails } from './utils.ts/types';
+import { CreateUserDto } from './dto';
  
 @Injectable()
 export class UserService {
@@ -12,21 +11,17 @@ export class UserService {
 		private userRepository: Repository<User>,
 		) {}
 
-	async findByID(id: number) {
-		const user = await this.userRepository.findOne({ id });
-		if (user) {
-		  return user;
-		}
-		throw new NotFoundException('User not found'); 
+	async findByID(id: number): Promise<User | undefined> {
+		return await this.userRepository.findOne({ id });
 	}
 
-	async findByIntraID(intraID: string) {
+	async findByIntraID(intraID: string): Promise<User | undefined> {
 		return await this.userRepository.findOne({ intraID });
 	}
 
-	createUser(userData: CreateUserDto) {
+	async createUser(userData: CreateUserDto) {
 		const newUser = this.userRepository.create(userData);
-		return this.userRepository.save(newUser);
+		return await this.userRepository.save(newUser);
 	}
 
 }
