--- backend/src/user/utils.ts/types.ts
@@ -1,8 +1,0 @@
-import { User } from "../user.entity"
-
-export type UserDetails = {
-	intraID: string;
-	username: string;
-};
-
-export type Done = (err: Error, user: User) => void;
